Updated Readme & dry-run code. -shuffled around runCommand and dry-run functionality to the main throttler.go -implemented dry-run for ipfw
