+++ throttler/tc.go
@@ -1,10 +1,7 @@
 package throttler
 
 import (
-	"bufio"
-	"errors"
 	"fmt"
-	"os/exec"
 	"strconv"
 	"strings"
 )
@@ -35,8 +32,6 @@ const (
 )
 
 type tcThrottler struct{}
-
-var DRY bool
 
 func (t *tcThrottler) setup(c *Config) error {
 
@@ -243,45 +238,4 @@ func (t *tcThrottler) exists(c *Config) bool {
 
 func (t *tcThrottler) check() string {
 	return tcCheck
-}
-
-func runCommand(cmd string) error {
-	fmt.Println(cmd)
-	if !DRY {
-		err := exec.Command("/bin/sh", "-c", cmd).Run()
-		return err
-	}
-	return nil
-}
-
-func runCommandGetLines(cmd string) ([]string, error) {
-
-	lines := []string{}
-	child := exec.Command("/bin/sh", "-c", cmd)
-
-	out, err := child.StdoutPipe()
-	if err != nil {
-		return []string{}, err
-	}
-
-	err = child.Start()
-	if err != nil {
-		return []string{}, err
-	}
-
-	scanner := bufio.NewScanner(out)
-
-	for scanner.Scan() {
-		lines = append(lines, scanner.Text())
-	}
-	if err := scanner.Err(); err != nil {
-		return []string{}, errors.New(fmt.Sprint("Error reading standard input:", err))
-	}
-
-	err = child.Wait()
-	if err != nil {
-		return []string{}, err
-	}
-
-	return lines, nil
 }
